refactor: 💡 move privacy screen to folder
