+++ src/web/modules/privacy/screens/PrivacyScreen.tsx
@@ -1,10 +1,14 @@
-import { ROUTES } from '@common/modules/router/constants/common'
 import React, { useCallback } from 'react'
-import useNavigation from '@common/hooks/useNavigation'
+
 import BackButton from '@common/components/BackButton'
+import useNavigation from '@common/hooks/useNavigation'
+import { ROUTES } from '@common/modules/router/constants/common'
+
+import usePrivacyForm from '../hooks'
 
 const PrivacyScreen = () => {
   const { navigate } = useNavigation()
+  const { sdkState } = usePrivacyForm()
 
   const onBack = useCallback(() => {
     navigate(ROUTES.dashboard)
@@ -13,7 +17,7 @@ const PrivacyScreen = () => {
   return (
     <div>
       <BackButton onPress={onBack} />
-      Privacy
+      <h1>Privacy</h1>
     </div>
   )
 }
